requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cfgv==3.2.0
chardet==4.0.0
distlib==0.3.1
Django==3.1.5
djangorestframework==3.12.2
filelock==3.0.12
identify==1.5.12
idna==2.10
nodeenv==1.5.0
pre-commit==2.9.3
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
virtualenv==20.3.1
